# docs/admin-openapi.yaml
openapi: 3.0.3
info:
  title: 只因梗懂你 · 管理端 API
  description: |
    管理端接口（`/admin/**`）及登录、AI 运维相关接口。
    成功响应统一为 `{ code: 1, message, data }`；业务失败抛 BizException，常见 code：0/400/401/403/404。
    鉴权：`Authorization: Bearer <accessToken>`；`/admin/**` 要求 `ROLE_ADMIN`。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地后端
...
paths:
...
  /admin/users:
    get:
      tags: [AdminUser]
      summary: 分页查询用户
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: integer
            enum: [0, 1]
        - name: role
          in: query
          schema:
            type: string
            enum: [ROLE_USER, ROLE_ADMIN]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminUserPage'
  /admin/users/{id}/status:
    patch:
      tags: [AdminUser]
      summary: 更新用户状态
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: integer
                  enum: [0, 1]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminUser'
  /admin/users/{id}/role:
    patch:
      tags: [AdminUser]
      summary: 更新用户角色
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [ROLE_USER, ROLE_ADMIN]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminUser'
...
  /admin/sensitive-words:
    get:
      tags: [AdminSensitiveWord]
      summary: 分页查询敏感词
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: keyword
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
            enum: [POLITICS, PORN, ABUSE, AD, CUSTOM]
        - name: status
          in: query
          schema:
            type: integer
            enum: [0, 1]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminSensitiveWordPage'
    post:
      tags: [AdminSensitiveWord]
      summary: 新增敏感词
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSensitiveWordSaveRequestDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminSensitiveWord'
  /admin/sensitive-words/{id}:
    put:
      tags: [AdminSensitiveWord]
      summary: 更新敏感词
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSensitiveWordSaveRequestDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminSensitiveWord'
    delete:
      tags: [AdminSensitiveWord]
      summary: 删除敏感词
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
  /admin/sensitive-words/{id}/status:
    patch:
      tags: [AdminSensitiveWord]
      summary: 启停敏感词
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: integer
                  enum: [0, 1]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminSensitiveWord'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
        maximum: 50
    pathId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ResultBase:
      type: object
      properties:
        code:
          type: integer
          example: 1
        message:
          type: string
        data: {}
...
    PageMeta:
      type: object
      properties:
        list:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
...
    AdminUserListItemVO:
      type: object
      properties:
        id: { type: string }
        nickname: { type: string }
        avatar: { type: string }
        signature: { type: string }
        role: { type: string }
        status: { type: integer }
        createTime: { type: string }
        updateTime: { type: string }
...
    AdminSensitiveWordVO:
      type: object
      properties:
        id: { type: string }
        word: { type: string }
        category: { type: string }
        actionType: { type: integer }
        status: { type: integer }
        creator: { type: string }
        createTime: { type: string }
        updateTime: { type: string }
    AdminSensitiveWordSaveRequestDTO:
      type: object
      required: [word, category, actionType]
      properties:
        word: { type: string }
        category: { type: string, enum: [POLITICS, PORN, ABUSE, AD, CUSTOM] }
        actionType: { type: integer, enum: [1, 2, 3] }
        status: { type: integer, enum: [0, 1] }
...
    ResultAdminUserPage:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    list:
                      type: array
                      items:
                        $ref: '#/components/schemas/AdminUserListItemVO'
    ResultAdminUser:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminUserListItemVO'
...
    ResultAdminSensitiveWordPage:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    list:
                      type: array
                      items:
                        $ref: '#/components/schemas/AdminSensitiveWordVO'
    ResultAdminSensitiveWord:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminSensitiveWordVO'
...